# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Tails-FinderBD",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "A RESTful API for managing animal adverts. Supports user registration, authentication, managing adverts, filters, and enum options.\n"
  },
  "servers": [
    {
      "url": "https://tails-finderbd.onrender.com/api",
      "description": "Production server"
    },
    {
      "url": "http://127.0.0.1:3000/api",
      "description": "Local server"
    }
  ],
...
  "paths": {
...
    "/adverts": {
      "get": {
        "operationId": "getAdvertsService",
        "summary": "Retrieve all animal ads with filters, pagination and sorting",
        "tags": [
          "AnimalAds"
        ],
        "description": "Returns a paginated array of animal ads. Supports filtering by archived status, advert status, species, colors, city, and district. Results can also be sorted by specific fields. If no ads are found, a 404 error is returned.",
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 20
            },
            "description": "Number of results per page"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "updatedAt"
              ],
              "default": "createdAt"
            },
            "description": "Field to sort results by"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "description": "Sort order (ascending or descending)"
          },
          {
            "in": "query",
            "name": "archived",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by archived status (true/false)"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "found",
                "lost"
              ]
            },
            "description": "Filter by advert status (found/lost)"
          },
          {
            "in": "query",
            "name": "species",
            "schema": {
              "type": "string",
              "example": "dog"
            },
            "description": "Filter by animal species"
          },
          {
            "in": "query",
            "name": "colors",
            "schema": {
              "type": "string",
              "example": "black,white"
            },
            "description": "Filter by animal colors. Multiple colors can be passed, separated by commas."
          },
          {
            "in": "query",
            "name": "city",
            "schema": {
              "type": "string",
              "example": "Київ"
            },
            "description": "Filter by city"
          },
          {
            "in": "query",
            "name": "district",
            "schema": {
              "type": "string",
              "example": "Оболонський"
            },
            "description": "Filter by district"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the list of animal ads",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/advert"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "example": 120,
                          "description": "Total number of adverts matching filters"
                        },
                        "page": {
                          "type": "integer",
                          "example": 1,
                          "description": "Current page number"
                        },
                        "perPage": {
                          "type": "integer",
                          "example": 20,
                          "description": "Number of results per page"
                        },
                        "totalPages": {
                          "type": "integer",
                          "example": 6,
                          "description": "Total number of pages"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No animal ads found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "No animal ads found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error: Database connection failed"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createAdvertService",
        "summary": "Create a new animal advert",
        "tags": [
          "AnimalAds"
        ],
        "description": "Creates a new animal advert for a found or lost animal. Requires authentication. Photos are uploaded as files. After creation, the advert is linked to the user and possible matches may trigger notifications to other users.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "photos",
                  "status",
                  "animal",
                  "context"
                ],
                "properties": {
                  "photos": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "One or more photos of the animal"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "blackFur",
                      "whiteEar"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "found",
                      "lost"
                    ],
                    "example": "lost"
                  },
                  "animal": {
                    "type": "object",
                    "required": [
                      "species",
                      "colors"
                    ],
                    "properties": {
                      "species": {
                        "type": "string",
                        "example": "dog"
                      },
                      "breed": {
                        "type": "string",
                        "example": "labrador"
                      },
                      "colors": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "example": "black"
                        }
                      },
                      "sex": {
                        "type": "string",
                        "enum": [
                          "самець",
                          "самка",
                          "інше"
                        ],
                        "example": "самка"
                      },
                      "age": {
                        "type": "string",
                        "example": "2 роки"
                      },
                      "size": {
                        "type": "string",
                        "enum": [
                          "малий",
                          "середній",
                          "великий",
                          "інше"
                        ],
                        "example": "середній"
                      },
                      "features": {
                        "type": "string",
                        "example": "Має червону хустку"
                      }
                    }
                  },
                  "context": {
                    "type": "object",
                    "required": [
                      "location",
                      "date"
                    ],
                    "properties": {
                      "location": {
                        "type": "object",
                        "required": [
                          "type",
                          "coordinates",
                          "city"
                        ],
                        "properties": {
                          "type": {
                            "type": "string",
                            "enum": [
                              "Point"
                            ],
                            "example": "Point"
                          },
                          "coordinates": {
                            "type": "array",
                            "description": "[longitude, latitude]",
                            "items": {
                              "type": "number"
                            },
                            "example": [
                              30.5234,
                              50.4501
                            ]
                          },
                          "city": {
                            "type": "string",
                            "example": "Київ"
                          },
                          "district": {
                            "type": "string",
                            "example": "Оболонський"
                          },
                          "address": {
                            "type": "string",
                            "example": "вул. Грушевського, 1"
                          }
                        }
                      },
                      "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01T12:00:00.000Z"
                      },
                      "description": {
                        "type": "string",
                        "maxLength": 1000,
                        "example": "Собака загубився біля парку. Дуже дружній."
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Advert successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Advert successfully created!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/advert"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid advert data"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - user not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error: Database connection failed"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/adverts/{id}": {
      "get": {
        "operationId": "getAdvertByIdService",
        "summary": "Retrieve an animal ad by ID",
        "tags": [
          "AnimalAds"
        ],
        "description": "Returns a single animal ad by its ID. If the ID is invalid, a 400 error is returned. If the advert does not exist, a 404 error is returned.",
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "665d14a36a70b4794fbcab12"
            },
            "description": "The ID of the advert to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the advert",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/advert"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid advert ID format",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid advert ID format"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Advert not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Advert not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error: Database connection failed"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateAdvertService",
        "summary": "Update an existing animal advert",
        "tags": [
          "AnimalAds"
        ],
        "description": "Updates an existing animal advert by its ID. Only the provided fields will be updated. Supports updating photos (multipart/form-data). Requires authentication.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "665d14a36a70b4794fbcab12"
            },
            "description": "The ID of the advert to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "photos": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "One or more new photos of the animal"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "blackFur",
                      "whiteEar"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "found",
                      "lost"
                    ],
                    "example": "found"
                  },
                  "animal": {
                    "type": "object",
                    "properties": {
                      "species": {
                        "type": "string",
                        "example": "cat"
                      },
                      "breed": {
                        "type": "string",
                        "example": "siamese"
                      },
                      "colors": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "example": "white"
                        }
                      },
                      "sex": {
                        "type": "string",
                        "enum": [
                          "самець",
                          "самка",
                          "інше"
                        ],
                        "example": "самець"
                      },
                      "age": {
                        "type": "string",
                        "example": "3 роки"
                      },
                      "size": {
                        "type": "string",
                        "enum": [
                          "малий",
                          "середній",
                          "великий",
                          "інше"
                        ],
                        "example": "малий"
                      },
                      "features": {
                        "type": "string",
                        "example": "Має зелену нашийну стрічку"
                      }
                    }
                  },
                  "context": {
                    "type": "object",
                    "properties": {
                      "location": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "enum": [
                              "Point"
                            ],
                            "example": "Point"
                          },
                          "coordinates": {
                            "type": "array",
                            "description": "[longitude, latitude]",
                            "items": {
                              "type": "number"
                            },
                            "example": [
                              30.5234,
                              50.4501
                            ]
                          },
                          "city": {
                            "type": "string",
                            "example": "Київ"
                          },
                          "district": {
                            "type": "string",
                            "example": "Шевченківський"
                          },
                          "address": {
                            "type": "string",
                            "example": "вул. Хрещатик, 12"
                          }
                        }
                      },
                      "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-01T12:00:00.000Z"
                      },
                      "description": {
                        "type": "string",
                        "maxLength": 1000,
                        "example": "Кота бачили біля магазину."
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Advert successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Advert successfully updated!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/advert"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid advert ID format or bad request data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid advert ID format"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - user not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Advert not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Advert not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error: Database connection failed"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "advert": {
        "type": "object",
        "required": [
          "_id",
          "photos",
          "status",
          "animal",
          "context",
          "user"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "example": "665d14a36a70b4794fbcab12"
          },
          "photos": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "https://example.com/uploads/dog1.jpg"
            },
            "description": "Масив URL фотографій (галерея)"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "friendly"
            },
            "description": "Додаткові мітки"
          },
          "status": {
            "type": "string",
            "enum": [
              "found",
              "lost"
            ],
            "example": "lost",
            "description": "Статус оголошення (знайдено / загублено)"
          },
          "animal": {
            "type": "object",
            "required": [
              "species",
              "colors"
            ],
            "properties": {
              "species": {
                "type": "string",
                "example": "dog"
              },
              "breed": {
                "type": "string",
                "example": "labrador"
              },
              "colors": {
                "type": "array",
                "items": {
                  "type": "string",
                  "example": "black"
                }
              },
              "sex": {
                "type": "string",
                "enum": [
                  "самець",
                  "самка",
                  "інше"
                ],
                "example": "самка"
              },
              "age": {
                "type": "string",
                "example": "2 роки"
              },
              "size": {
                "type": "string",
                "enum": [
                  "малий",
                  "середній",
                  "великий",
                  "інше"
                ],
                "example": "середній"
              },
              "features": {
                "type": "string",
                "example": "Має червону нашийну хустку"
              }
            }
          },
          "context": {
            "type": "object",
            "required": [
              "location",
              "date"
            ],
            "properties": {
              "location": {
                "type": "object",
                "required": [
                  "type",
                  "coordinates",
                  "city"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "Point"
                    ],
                    "example": "Point"
                  },
                  "coordinates": {
                    "type": "array",
                    "description": "[longitude, latitude]",
                    "items": {
                      "type": "number"
                    },
                    "example": [
                      30.5234,
                      50.4501
                    ]
                  },
                  "city": {
                    "type": "string",
                    "example": "Київ"
                  },
                  "district": {
                    "type": "string",
                    "example": "Оболонський"
                  },
                  "address": {
                    "type": "string",
                    "example": "вул. Грушевського, 1"
                  }
                }
              },
              "date": {
                "type": "string",
                "format": "date-time",
                "example": "2024-06-01T12:00:00.000Z"
              },
              "description": {
                "type": "string",
                "maxLength": 1000,
                "example": "Собака загубився біля парку. Дуже дружній."
              }
            }
          },
          "user": {
            "type": "string",
            "example": "665d14a36a70b4794fbcab34",
            "description": "ID користувача, який створив оголошення"
          },
          "archived": {
            "type": "boolean",
            "default": false,
            "example": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-06-01T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-06-01T12:30:00.000Z"
          }
        }
      }
    },
...
  }
}